sql/parser: move ColumnType definitions into separate files
